requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
google-generativeai==0.3.0
beautifulsoup4==4.12.3
requests==2.31.0
gunicorn==22.0.0
python-dotenv==1.0.1
